backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
httpx==0.28.0
pyjwt==2.10.1
python-dotenv==1.0.1
pydantic==2.10.3
